+++ src/imio/history/safe_utils.py
@@ -2,6 +2,7 @@
 # flake8: noqa
 
 from imio.history.utils import get_all_history_attr
+from imio.history.utils import get_last_event
 from imio.history.utils import getLastAction
 from imio.history.utils import getLastWFAction
 from imio.history.utils import getPreviousEvent